\name{combined_scaleanalysis}
\alias{combined_scaleanalysis}
\title{Determine standard values for scale analyses}
\usage{
combined_scaleanalysis(
    data, 
    round = 2
)
}
\description{
This functions allows quickly determinining standard values for scale analyses including Cronbach's alpha, Omega, and the confidence intervals for alpha. 
}

\arguments{
\item{data}{data input, as data frame or matrix}
\item{round}{rounding factor}
}
\examples{}
\references{
Revelle, W. (2021) psych: Procedures for Personality and Psychological Research, Northwestern University, Evanston, Illinois, USA, https://CRAN.R-project.org/package=psych Version = 2.1.9.
}
