\name{scales}
\alias{scales}
\title{Determine standard values for scale analyses}
\usage{
scales(
    data, 
    round = 2
)
}
\description{
This function determines standard scale analysis values, including Cronbach's alpha, Omega, and confidence intervals, as well as the greatest lower bound value for the data.}

\arguments{
\item{data}{data input, as data frame or matrix}
\item{round}{rounding factor}
}
\examples{}
\references{
Revelle, W. (2021) psych: Procedures for Personality and Psychological Research, Northwestern University, Evanston, Illinois, USA, https://CRAN.R-project.org/package=psych Version = 2.1.9.
}
